### 2018-06-21 #229 @noamross
Thank you for your submission, @zkamvar!  Can you comment on how your package fits into our Aims and Scope? (https://ropensci.github.io/dev_guide/policies.html#aims-and-scope).  This is an edge case that might go either way - on one hand, I think this falls under a statistical method, which is outside our scope.  However, it's possible it might make sense under the reproducibility category.
