### 2018-06-21 #229 @zkamvar
Hi @noamross,

Yes, the question of Aims and Scope is a bit tricky for this one.

I believe this fits under reproducibility because the goal of this package is to provide researchers with a tool that aids in selecting positive controls for reproducibility in population genetic studies. Otherwise, this process of selecting representative samples for re-genotyping is normally performed by hand. Thus, while a bit niche, this tool provides a way of tracking the provenance for the choice of the positive controls.
