### 2018-06-27 #229 @noamross
Sorry for the delay on this.  The editorial board discussed this and decided that, while this is an edge case, if submitted we'd consider it in scope.  We would likely aim to have at least one reviewer from a non-genetic field to consider other applications and associated documentation - you may want to consider if there are other relevant examples to include.
